% Generated by roxygen2 (4.0.1): do not edit by hand
\name{blueback_run_settings}
\alias{blueback_run_settings}
\title{function to return parameter settings for herring_all_analysis() for blueback}
\usage{
blueback_run_settings()
}
\description{
this returns parameter settings as a list so you can say do.call(herring_all_analyses, args = blueback_run_settings())
}

