% Generated by roxygen2 (4.0.1): do not edit by hand
\name{brewerRamp}
\alias{brewerRamp}
\title{create a ramp of colors starting from the last and proceeding to the y-th of a color brewer palette}
\usage{
brewerRamp(brew.pal, n = 7, y = 3)
}
\arguments{
\item{brew.pal}{The color brewer palette.  By name.}

\item{n}{The number of colors to return in the ramp}

\item{y}{Endpoint on the "light" side of the palette}
}
\description{
create a ramp of colors starting from the last and proceeding to the y-th of a color brewer palette
}
\examples{
rr <- brewerRamp("Purples", 9)
image(1:length(rr), 1, as.matrix(1:length(rr)), col = rr, ylab = "", xaxt = "n", yaxt = "n", bty = "n")
}

