% Generated by roxygen2 (4.0.1): do not edit by hand
\name{do_mixture_analysis}
\alias{do_mixture_analysis}
\title{analyzes a mixture from bycatch data frame to the end}
\usage{
do_mixture_analysis(bycatch_df, bycatch_locus_columns, baseline_df,
  rep_unit_path)
}
\description{
analyzes a mixture from bycatch data frame to the end
}

