% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/herring-main-func.R
\name{herring_all_analyses}
\alias{herring_all_analyses}
\title{main wrapper for doing all the analyses}
\usage{
herring_all_analyses(baseline_path, lat_long_path, rep_unit_path, locus_columns,
  bycatch_path, bycatch_locus_columns)
}
\description{
main wrapper for doing all the analyses
}

