% Generated by roxygen2 (4.0.1): do not edit by hand
\name{make_baseline_file}
\alias{make_baseline_file}
\title{make a gsi-sim file from a gpiper data frame}
\usage{
make_baseline_file(baseline_df)
}
\arguments{
\item{baseline.df}{a data frame of the sort returned by \code{\link{herring_csv2gpiper}}}
}
\value{
A factor vector of the populations in the baseline.  Has a side effect of writing a file named
gsi_sim_file.txt in the current working directory.
}
\description{
make a gsi-sim file from a gpiper data frame
}

