% Generated by roxygen2 (4.0.1): do not edit by hand
\name{read_and_stratify_bycatch}
\alias{read_and_stratify_bycatch}
\title{reads the bycatch into a data frame and makes each stratum a data frame in a list}
\usage{
read_and_stratify_bycatch(bycatch_path, dropBelow = 5,
  stratifyOn = c("Year", "Season", "Region", "Target.Fishery", "Gear.Type"),
  bycatch_locus_columns)
}
\arguments{
\item{bycatch_path}{The path to the bycatch file.}

\item{dropBelow}{Any stratum with fewer than dropBelow fish will be dropped from consideration}

\item{stratifyOn}{A character vector naming the columns holding the factors that determine the strata to
stratify the data into}

\item{bycatch_locus_columns}{integer vector with the columns in the original CSV \emph{bycatch} file that have the loci
that you want to use.  Should be twice as many columns as loci.}
}
\description{
reads the bycatch into a data frame and makes each stratum a data frame in a list
}

