% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simple_barplot}
\alias{simple_barplot}
\title{make a simple barplot of proportions assigned from an assignment array}
\usage{
simple_barplot(ass_array, cols = c("red", "blue", "green", "orange"))
}
\arguments{
\item{ass_array}{The self assignment arrray, as you might find as one of the
components returned by gsi_self_assignment}

\item{cols}{the colors to use}
}
\description{
make a simple barplot of proportions assigned from an assignment array
}

