% Generated by roxygen2 (4.0.0): do not edit by hand
\name{to_pops_barplot}
\alias{to_pops_barplot}
\title{make a barplot that has different densities of the same color for pops within the same reporting unit}
\usage{
to_pops_barplot(ass_array, rep_groups, brew.pals = c("Reds", "Blues",
  "Greens"))
}
\arguments{
\item{ass_array}{An assignment array with rows being from and cols being "to". Where individuals
are assigned to populations.}

\item{rep_groups}{a factor vector given the reporting groups each populations belongs to.  For this to
work, the pops must all be ordered so that they are all next to one another within reporting groups}

\item{brew.pals}{color brewer palette names.  There should be one for every reporting unit.  Will recycle as
necessary.}
}
\description{
make a barplot that has different densities of the same color for pops within the same reporting unit
}

