% Generated by roxygen2 (4.0.1): do not edit by hand
\name{write_gsi_sim_mixfile_from_bycatch_data_frame}
\alias{write_gsi_sim_mixfile_from_bycatch_data_frame}
\title{writes a gsi_sim mixture file from a data frame}
\usage{
write_gsi_sim_mixfile_from_bycatch_data_frame(df, bycatch_locus_columns)
}
\description{
writes a gsi_sim mixture file from a data frame
}

